PASS-2: Move env to base of package
